requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
groq
sentence-transformers
faiss-cpu
torch
python-dotenv